verify-ppb/pharmtechs/requirements-dev.txt
# Development and testing dependencies
-r requirements.txt

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
requests-mock==1.11.0
pytest-flask==1.3.0

# Code quality
black==24.3.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0
types-requests==2.31.0.20240125

# Pre-commit hooks
pre-commit==3.6.0

# Development tools
python-dotenv==1.0.1
ipython==8.21.0
